requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.8.0
async-timeout==4.0.3
attrs==24.2.0
blinker==1.9.0
Brotli==1.1.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
cssselect2==0.7.0
dnspython==2.7.0
eval_type_backport==0.2.0
exceptiongroup==1.2.2
filelock==3.16.1
Flask==3.1.0
fonttools==4.54.1
frozenlist==1.5.0
google-api-core==2.24.0
google-api-python-client==2.159.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.66.0
gspread==6.1.4
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.3
Markdown==3.7
markdown-it-py==3.0.0
markdown2==2.4.12
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.1.0
numpy==2.0.2
oauth2client==4.1.3
oauthlib==3.2.2
packaging==24.2
pandas==2.2.3
pdfkit==1.0.0
pillow==10.4.0
propcache==0.2.0
proto-plus==1.25.0
protobuf==5.29.3
pyarrow==18.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydyf==0.11.0
Pygments==2.18.0
pymongo==4.6.1
PyMuPDF==1.24.11
pyparsing==3.2.1
PyPDF2==3.0.1
pyphen==0.17.0
pyTelegramBotAPI==4.4.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-markdown-math==0.8
python-telegram-bot==21.10
pytz==2024.2
redis==5.0.1
reportlab==4.2.5
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rsa==4.9
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
tabulate==0.9.0
tinycss2==1.4.0
together==1.3.2
tqdm==4.67.0
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
weasyprint==60.2
webencodings==0.5.1
Werkzeug==3.1.3
yarl==1.17.1
zipp==3.21.0
zopfli==0.2.3.post1
